+++ service.go
@@ -6,12 +6,12 @@ import (
 
 // PrimeFinder finds primes in a specified range and calculates their sum.
 type PrimeService interface {
-	PrimeSum(uint64, bool, uint8) (uint64, []uint64, error)
+	PrimeSum(uint64, bool, uint64) (uint64, []uint64, error)
 }
 
 type primeService struct{}
 
-func (primeService) PrimeSum(n uint64, lst bool, nr uint8) (uint64, []uint64, error) {
+func (primeService) PrimeSum(n uint64, lst bool, nr uint64) (uint64, []uint64, error) {
 	sum, primes, err := primes.PrimeSum(n, lst, nr)
 	return sum, primes, err
 }
